requirements.txt
psycopg2-binary==2.9.7
Flask==2.3.2
flask_cors
python-dotenv
yfinance